实现观看历史记录管理，继续播放等

+++ app/player/src/main/java/com/zyl315/player/player/VideoView.java
@@ -641,7 +641,7 @@ public class VideoView<P extends AbstractPlayer> extends FrameLayout
     /**
      * 一开始播放就seek到预先设置好的位置
      */
-    public void skipPositionWhenPlay(int position) {
+    public void skipPositionWhenPlay(long position) {
         this.mCurrentPosition = position;
     }
